% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/berate-package.r
\docType{package}
\name{berate}
\alias{berate}
\alias{berate-package}
\title{Provide Insightful Motivation During Interactive Sessions}
\description{
Provide Insightful Motivation During Interactive Sessions
}
\author{
Bob Rudis (@hrbrmstr)
}

